requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Production WSGI Server
gunicorn==21.2.0
waitress==2.1.2

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.38.0

# Computer Vision
opencv-python-headless==4.8.1.78
pillow>=10.0.0
facenet-pytorch

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
google-generativeai
requests>=2.31.0

# Environment & Config
python-dotenv==1.0.0

# Security
cryptography>=41.0.0

# Optional (if you want advanced GNN libs):
# torch-geometric (installation varies by OS) or dgl